requirements.txt
pandas==2.0.3
streamlit==1.29.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3
matplotlib==3.8.2
plotly==5.18.0
wordcloud==1.9.3
python-dotenv==1.0.0
mlflow==2.9.2
underthesea==6.7.0
regex==2023.10.3
requests==2.31.0
